+++ docs/data-sources/graph_beta_identity_and_access_directory_role.md
@@ -1,0 +1,48 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "microsoft365_graph_beta_identity_and_access_directory_role Data Source - Microsoft 365"
+subcategory: ""
+description: |-
+  Retrieves activated directory roles from Microsoft Entra ID using the /directoryRoles endpoint. Returns tenant-specific directoryRole object IDs required by resources such as microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment. Supports lookup by role object ID, display name, or listing all activated roles.
+---
+
+# microsoft365_graph_beta_identity_and_access_directory_role (Data Source)
+
+Retrieves activated directory roles from Microsoft Entra ID using the `/directoryRoles` endpoint. Returns tenant-specific directoryRole object IDs required by resources such as `microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment`. Supports lookup by role object ID, display name, or listing all activated roles.
+
+
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Optional
+
+- `display_name` (String) Filter activated directory roles by display name (exact match, e.g. 'User Administrator'). Uses OData `$filter=displayName eq '...'`. Conflicts with `role_id` and `list_all`.
+- `list_all` (Boolean) Retrieve all activated directory roles in the tenant. Conflicts with `role_id` and `display_name`.
+- `role_id` (String) The tenant-specific object ID of the activated directory role. Conflicts with `display_name` and `list_all`.
+- `timeouts` (Attributes) (see [below for nested schema](#nestedatt--timeouts))
+
+### Read-Only
+
+- `items` (Attributes List) The list of activated directory roles that match the filter criteria. (see [below for nested schema](#nestedatt--items))
+
+<a id="nestedatt--timeouts"></a>
+### Nested Schema for `timeouts`
+
+Optional:
+
+- `create` (String) A string that can be [parsed as a duration](https://pkg.go.dev/time#ParseDuration) consisting of numbers and unit suffixes, such as "30s" or "2h45m". Valid time units are "s" (seconds), "m" (minutes), "h" (hours).
+- `delete` (String) A string that can be [parsed as a duration](https://pkg.go.dev/time#ParseDuration) consisting of numbers and unit suffixes, such as "30s" or "2h45m". Valid time units are "s" (seconds), "m" (minutes), "h" (hours). Setting a timeout for a Delete operation is only applicable if changes are saved into state before the destroy operation occurs.
+- `read` (String) A string that can be [parsed as a duration](https://pkg.go.dev/time#ParseDuration) consisting of numbers and unit suffixes, such as "30s" or "2h45m". Valid time units are "s" (seconds), "m" (minutes), "h" (hours). Read operations occur during any refresh or planning operation when refresh is enabled.
+- `update` (String) A string that can be [parsed as a duration](https://pkg.go.dev/time#ParseDuration) consisting of numbers and unit suffixes, such as "30s" or "2h45m". Valid time units are "s" (seconds), "m" (minutes), "h" (hours).
+
+
+<a id="nestedatt--items"></a>
+### Nested Schema for `items`
+
+Read-Only:
+
+- `description` (String) The description of the directory role.
+- `display_name` (String) The display name of the directory role (e.g. 'User Administrator').
+- `id` (String) The tenant-specific object ID of the activated directory role. e.g you can use this value as `directory_role_id` in `microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment`.
+- `role_template_id` (String) The well-known, cross-tenant role template ID. This is the same identifier across all tenants for the same built-in role (e.g. `fe930be7-5e62-47db-91af-98c3a49a38b1` for User Administrator).
