+++ examples/resources/microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment/import.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+# Import format: administrative_unit_id/scoped_role_membership_id
+terraform import microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment.example 00000000-0000-0000-0000-000000000000/11111111-1111-1111-1111-111111111111
