+++ internal/provider/resources.go
@@ -117,7 +117,7 @@ import (
 	// Graph Beta - Identity and Access resources
 	graphBetaIdentityAndAccessAdministrativeUnit "github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/resources/identity_and_access/graph_beta/administrative_unit"
 	graphBetaIdentityAndAccessAdministrativeUnitMembership "github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/resources/identity_and_access/graph_beta/administrative_unit_membership"
-	graphBetaIdentityAndAccessAdministrativeUnitRoleAssignment "github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/resources/identity_and_access/graph_beta/administrative_unit_role_assignment"
+	graphBetaIdentityAndAccessAdministrativeUnitDirectoryRoleAssignment "github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/resources/identity_and_access/graph_beta/administrative_unit_directory_role_assignment"
 	graphBetaIdentityAndAccessAttributeSet "github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/resources/identity_and_access/graph_beta/attribute_set"
 	graphBetaIdentityAndAccessAuthenticationContext "github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/resources/identity_and_access/graph_beta/authentication_context"
 	graphBetaIdentityAndAccessAuthenticationStrengthPolicy "github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/resources/identity_and_access/graph_beta/authentication_strength_policy"
@@ -302,7 +302,7 @@ func (p *M365Provider) Resources(ctx context.Context) []func() resource.Resource
 		// Graph Beta - Identity and Access resources
 		graphBetaIdentityAndAccessAdministrativeUnit.NewAdministrativeUnitResource,
 		graphBetaIdentityAndAccessAdministrativeUnitMembership.NewAdministrativeUnitMembershipResource,
-		graphBetaIdentityAndAccessAdministrativeUnitRoleAssignment.NewAdministrativeUnitRoleAssignmentResource,
+		graphBetaIdentityAndAccessAdministrativeUnitDirectoryRoleAssignment.NewAdministrativeUnitDirectoryRoleAssignmentResource,
 		graphBetaIdentityAndAccessAttributeSet.NewAttributeSetResource,
 		graphBetaIdentityAndAccessAuthenticationContext.NewAuthenticationContextResource,
 		graphBetaIdentityAndAccessAuthenticationStrengthPolicy.NewAuthenticationStrengthPolicyResource,
