--- internal/services/resources/identity_and_access/graph_beta/administrative_unit_role_assignment/crud.go
@@ -1,252 +1,0 @@
-package graphBetaAdministrativeUnitRoleAssignment
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	"github.com/deploymenttheory/terraform-provider-microsoft365/internal/constants"
-	"github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/common/crud"
-	errors "github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/common/errors/kiota"
-	sharedmodels "github.com/deploymenttheory/terraform-provider-microsoft365/internal/services/common/shared_models/graph_beta"
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
-)
-
-// Create handles the Create operation for administrative unit scoped role assignment resources.
-//
-// Operation: Assigns a directory role to a user scoped to a specific administrative unit
-// API Calls:
-//   - POST /administrativeUnits/{id}/scopedRoleMembers
-//
-// Reference: https://learn.microsoft.com/en-us/graph/api/administrativeunit-post-scopedrolemembers?view=graph-rest-beta
-func (r *AdministrativeUnitRoleAssignmentResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	var object AdministrativeUnitRoleAssignmentResourceModel
-
-	tflog.Debug(ctx, fmt.Sprintf("Starting creation of resource: %s", ResourceName))
-
-	resp.Diagnostics.Append(req.Plan.Get(ctx, &object)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	ctx, cancel := crud.HandleTimeout(ctx, object.Timeouts.Create, CreateTimeout*time.Second, &resp.Diagnostics)
-	if cancel == nil {
-		return
-	}
-	defer cancel()
-
-	if !validateRequest(ctx, r.client, &object, &resp.Diagnostics) {
-		return
-	}
-
-	requestBody, err := constructResource(ctx, &object)
-	if err != nil {
-		resp.Diagnostics.AddError(
-			"Error constructing resource",
-			fmt.Sprintf("Could not construct resource: %s: %s", ResourceName, err.Error()),
-		)
-		return
-	}
-
-	administrativeUnitID := object.AdministrativeUnitID.ValueString()
-
-	createdResource, err := r.client.
-		AdministrativeUnits().
-		ByAdministrativeUnitId(administrativeUnitID).
-		ScopedRoleMembers().
-		Post(ctx, requestBody, nil)
-
-	if err != nil {
-		errors.HandleKiotaGraphError(ctx, err, resp, constants.TfOperationCreate, r.WritePermissions)
-		return
-	}
-
-	if createdResource == nil {
-		resp.Diagnostics.AddError(
-			"Error creating resource",
-			fmt.Sprintf("Could not create %s. The API returned a nil response.", ResourceName),
-		)
-		return
-	}
-
-	if createdResource.GetId() == nil {
-		resp.Diagnostics.AddError(
-			"Error extracting resource ID",
-			fmt.Sprintf("Could not extract ID from created resource: %s. The API returned nil for the ID.", ResourceName),
-		)
-		return
-	}
-
-	object = MapRemoteResourceStateToTerraform(ctx, object, createdResource)
-
-	tflog.Debug(ctx, fmt.Sprintf("Successfully created %s with ID: %s", ResourceName, object.ID.ValueString()))
-
-	resp.Diagnostics.Append(resp.State.Set(ctx, &object)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	readReq := resource.ReadRequest{State: resp.State, ProviderMeta: req.ProviderMeta}
-	stateContainer := &crud.CreateResponseContainer{CreateResponse: resp}
-
-	opts := crud.DefaultReadWithRetryOptions()
-	opts.Operation = constants.TfOperationCreate
-	opts.ResourceTypeName = ResourceName
-
-	err = crud.ReadWithRetry(ctx, r.Read, readReq, stateContainer, opts)
-	if err != nil {
-		resp.Diagnostics.AddError(
-			"Error reading resource state after create",
-			fmt.Sprintf("Could not read resource state: %s: %s", ResourceName, err.Error()),
-		)
-		return
-	}
-
-	tflog.Debug(ctx, fmt.Sprintf("Finished Create Method: %s", ResourceName))
-}
-
-// Read handles the Read operation for administrative unit scoped role assignment resources.
-//
-// Operation: Retrieves a scoped role membership by its ID within an administrative unit
-// API Calls:
-//   - GET /administrativeUnits/{id}/scopedRoleMembers/{scopedRoleMembershipId}
-//
-// Reference: https://learn.microsoft.com/en-us/graph/api/administrativeunit-get-scopedrolemembers?view=graph-rest-beta
-func (r *AdministrativeUnitRoleAssignmentResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var object AdministrativeUnitRoleAssignmentResourceModel
-	var identity sharedmodels.ResourceIdentity
-
-	tflog.Debug(ctx, fmt.Sprintf("Starting Read method for: %s", ResourceName))
-
-	operation := constants.TfOperationRead
-	if ctxOp := ctx.Value("retry_operation"); ctxOp != nil {
-		if opStr, ok := ctxOp.(string); ok {
-			operation = opStr
-		}
-	}
-
-	resp.Diagnostics.Append(req.State.Get(ctx, &object)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	administrativeUnitID := object.AdministrativeUnitID.ValueString()
-	scopedRoleMembershipID := object.ID.ValueString()
-
-	tflog.Debug(ctx, fmt.Sprintf("Reading %s with ID: %s for administrative unit: %s", ResourceName, scopedRoleMembershipID, administrativeUnitID))
-
-	ctx, cancel := crud.HandleTimeout(ctx, object.Timeouts.Read, ReadTimeout*time.Second, &resp.Diagnostics)
-	if cancel == nil {
-		return
-	}
-	defer cancel()
-
-	remoteResource, err := r.client.
-		AdministrativeUnits().
-		ByAdministrativeUnitId(administrativeUnitID).
-		ScopedRoleMembers().
-		ByScopedRoleMembershipId(scopedRoleMembershipID).
-		Get(ctx, nil)
-
-	if err != nil {
-		errors.HandleKiotaGraphError(ctx, err, resp, operation, r.ReadPermissions)
-		return
-	}
-
-	object = MapRemoteResourceStateToTerraform(ctx, object, remoteResource)
-
-	resp.Diagnostics.Append(resp.State.Set(ctx, &object)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	identity.ID = object.ID.ValueString()
-
-	if resp.Identity != nil {
-		resp.Diagnostics.Append(resp.Identity.Set(ctx, identity)...)
-		if resp.Diagnostics.HasError() {
-			return
-		}
-	}
-
-	tflog.Debug(ctx, fmt.Sprintf("Finished Read Method: %s", ResourceName))
-}
-
-// Update handles the Update operation for administrative unit scoped role assignment resources.
-//
-// Note: All fields on this resource require replacement, so Update should never be called in practice.
-// If called, it performs a read to refresh computed state.
-func (r *AdministrativeUnitRoleAssignmentResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var plan AdministrativeUnitRoleAssignmentResourceModel
-
-	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	readReq := resource.ReadRequest{State: resp.State, ProviderMeta: req.ProviderMeta}
-	stateContainer := &crud.UpdateResponseContainer{UpdateResponse: resp}
-
-	opts := crud.DefaultReadWithRetryOptions()
-	opts.Operation = constants.TfOperationUpdate
-	opts.ResourceTypeName = ResourceName
-
-	err := crud.ReadWithRetry(ctx, r.Read, readReq, stateContainer, opts)
-	if err != nil {
-		resp.Diagnostics.AddError(
-			"Error reading resource state after update",
-			fmt.Sprintf("Could not read resource state: %s: %s", ResourceName, err.Error()),
-		)
-		return
-	}
-
-	tflog.Debug(ctx, fmt.Sprintf("Finished Update Method: %s", ResourceName))
-}
-
-// Delete handles the Delete operation for administrative unit scoped role assignment resources.
-//
-// Operation: Removes a scoped role assignment from an administrative unit
-// API Calls:
-//   - DELETE /administrativeUnits/{id}/scopedRoleMembers/{scopedRoleMembershipId}
-//
-// Reference: https://learn.microsoft.com/en-us/graph/api/administrativeunit-delete-scopedrolemembers?view=graph-rest-beta
-func (r *AdministrativeUnitRoleAssignmentResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	var object AdministrativeUnitRoleAssignmentResourceModel
-
-	tflog.Debug(ctx, fmt.Sprintf("Starting deletion of resource: %s", ResourceName))
-
-	resp.Diagnostics.Append(req.State.Get(ctx, &object)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	ctx, cancel := crud.HandleTimeout(ctx, object.Timeouts.Delete, DeleteTimeout*time.Second, &resp.Diagnostics)
-	if cancel == nil {
-		return
-	}
-	defer cancel()
-
-	administrativeUnitID := object.AdministrativeUnitID.ValueString()
-	scopedRoleMembershipID := object.ID.ValueString()
-
-	tflog.Debug(ctx, fmt.Sprintf("Deleting scoped role membership %s from administrative unit %s", scopedRoleMembershipID, administrativeUnitID))
-
-	err := r.client.
-		AdministrativeUnits().
-		ByAdministrativeUnitId(administrativeUnitID).
-		ScopedRoleMembers().
-		ByScopedRoleMembershipId(scopedRoleMembershipID).
-		Delete(ctx, nil)
-
-	if err != nil {
-		errors.HandleKiotaGraphError(ctx, err, resp, constants.TfOperationDelete, r.WritePermissions)
-		return
-	}
-
-	tflog.Debug(ctx, fmt.Sprintf("Removing %s from Terraform state", ResourceName))
-
-	resp.State.RemoveResource(ctx)
-
-	tflog.Debug(ctx, fmt.Sprintf("Finished Delete Method: %s", ResourceName))
-}
