+++ templates/resources/graph_beta_identity_and_access_administrative_unit_directory_role_assignment.md.tmpl
@@ -1,0 +1,81 @@
+---
+page_title: "{{.Name}} {{.Type}} - {{.ProviderName}}"
+subcategory: "Identity and Access"
+description: |-
+{{ .Description | plainmarkdown | trimspace | prefixlines "  " }}
+---
+
+# {{.Name}} ({{.Type}})
+
+{{ .Description | trimspace }}
+
+## Microsoft Documentation
+
+- [Administrative units overview](https://learn.microsoft.com/en-us/entra/identity/role-based-access-control/administrative-units)
+- [scopedRoleMembership resource type](https://learn.microsoft.com/en-us/graph/api/resources/scopedrolemembership?view=graph-rest-beta)
+- [Add a scopedRoleMember](https://learn.microsoft.com/en-us/graph/api/administrativeunit-post-scopedrolemembers?view=graph-rest-beta)
+- [List scopedRoleMembers](https://learn.microsoft.com/en-us/graph/api/administrativeunit-list-scopedrolemembers?view=graph-rest-beta)
+- [Get a scopedRoleMember](https://learn.microsoft.com/en-us/graph/api/administrativeunit-get-scopedrolemembers?view=graph-rest-beta)
+- [Remove a scopedRoleMember](https://learn.microsoft.com/en-us/graph/api/administrativeunit-delete-scopedrolemembers?view=graph-rest-beta)
+
+## Microsoft Graph API Permissions
+
+The following client `application` permissions are needed in order to use this {{.Type | lower}}:
+
+**Required:**
+- `RoleManagement.ReadWrite.Directory`
+
+**Optional:**
+- `RoleManagement.Read.Directory`
+- `Directory.Read.All`
+
+## Version History
+
+| Version | Status | Notes |
+|---------|--------|-------|
+| v0.1.0-alpha | Experimental | Initial release |
+
+## Example Usage
+
+### AURA001: User Administrator Scoped to an Administrative Unit
+
+{{ tffile "examples/resources/microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment/resource_aura001_user_administrator.tf" }}
+
+### AURA002: Helpdesk Administrator Scoped to an Administrative Unit
+
+{{ tffile "examples/resources/microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment/resource_aura002_helpdesk_administrator.tf" }}
+
+### AURA003: Multiple Role Assignments on the Same Administrative Unit
+
+{{ tffile "examples/resources/microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment/resource_aura003_multiple_role_assignments.tf" }}
+
+### AURA004: Full Example — Create AU, User, and Role Assignment Together
+
+{{ tffile "examples/resources/microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment/resource_aura004_with_au_and_user.tf" }}
+
+{{ .SchemaMarkdown | trimspace }}
+
+## Important Notes
+
+### Immutability
+All input fields (`administrative_unit_id`, `directory_role_id`, `role_member_id`) are immutable. Any change destroys the existing assignment and creates a new one.
+
+### Providing the `directory_role_id`
+The `directory_role_id` requires the **tenant-specific directoryRole object ID**, not the well-known `roleTemplateId`. Use the [`microsoft365_graph_beta_identity_and_access_directory_role`](../data-sources/graph_beta_identity_and_access_directory_role) data source to look up the correct value by display name.
+
+### Supported Roles
+Not all directory roles support administrative unit scoping. For a full list see [roles that can be assigned with administrative unit scope](https://learn.microsoft.com/en-us/entra/identity/role-based-access-control/admin-units-assign-roles#roles-that-can-be-assigned-with-administrative-unit-scope).
+
+### Role Member Types
+The `role_member_id` must refer to a **user** object. Groups and service principals cannot be assigned to scoped roles.
+
+### Relationship with Other Resources
+- Use [`microsoft365_graph_beta_identity_and_access_administrative_unit`](../resources/graph_beta_identity_and_access_administrative_unit) to create the administrative unit.
+- Use [`microsoft365_graph_beta_identity_and_access_administrative_unit_membership`](../resources/graph_beta_identity_and_access_administrative_unit_membership) to manage users and groups that belong **to** the administrative unit.
+- This resource controls who has **administrative authority over** the administrative unit.
+
+## Import
+
+Import is supported using the following syntax:
+
+{{ codefile "shell" "examples/resources/microsoft365_graph_beta_identity_and_access_administrative_unit_directory_role_assignment/import.sh" }}
